refactor: HomeMenuをObjectからClassへ変更,homeパッケージに移動
