+++ src/main/scala/com/github/unchama/seichiassist/menus/TopLevelRouter.scala
@@ -4,12 +4,13 @@ import cats.effect.{IO, SyncIO}
 import com.github.unchama.menuinventory.LayoutPreparationContext
 import com.github.unchama.menuinventory.router.CanOpen
 import com.github.unchama.minecraft.actions.OnMinecraftServerThread
-import com.github.unchama.seichiassist.menus.HomeMenu.ConfirmationMenuEnvironment
+import com.github.unchama.seichiassist.menus.home.ConfirmationMenuEnvironment
 import com.github.unchama.seichiassist.menus.achievement.group.AchievementGroupMenu
 import com.github.unchama.seichiassist.menus.achievement.{
   AchievementCategoryMenu,
   AchievementMenu
 }
+import com.github.unchama.seichiassist.menus.home.{HomeMenu}
 import com.github.unchama.seichiassist.menus.minestack.{
   CategorizedMineStackMenu,
   MineStackMainMenu,
@@ -94,8 +95,7 @@ object TopLevelRouter {
     implicit lazy val achievementMenuEnv: AchievementMenu.Environment =
       new AchievementMenu.Environment
     implicit lazy val homeMenuEnv: HomeMenu.Environment = new HomeMenu.Environment
-    implicit lazy val homeConfirmationMenuEnv
-      : HomeMenu.ConfirmationMenuEnvironment.Environment =
+    implicit lazy val homeConfirmationMenuEnv: home.ConfirmationMenuEnvironment.Environment =
       new ConfirmationMenuEnvironment.Environment
     implicit lazy val achievementCategoryMenuEnv: AchievementCategoryMenu.Environment =
       new AchievementCategoryMenu.Environment
@@ -125,7 +125,7 @@ object TopLevelRouter {
       _.open
     implicit lazy val ioCanOpenAchievementGroupMenu: IO CanOpen AchievementGroupMenu = _.open
     implicit lazy val ioCanOpenHomeConfirmationMenu
-      : IO CanOpen HomeMenu.HomeChangeConfirmationMenu =
+      : IO CanOpen home.HomeChangeConfirmationMenu =
       _.open
     implicit lazy val ioCanOpenAchievementCategoryMenu: IO CanOpen AchievementCategoryMenu =
       _.open
@@ -141,8 +141,8 @@ object TopLevelRouter {
     implicit lazy val ioCanOpenRegionMenu: IO CanOpen RegionMenu.type = _.open
     implicit lazy val ioCanOpenActiveSkillMenu: IO CanOpen ActiveSkillMenu.type = _.open
     implicit lazy val ioCanOpenServerSwitchMenu: IO CanOpen ServerSwitchMenu.type = _.open
-    implicit lazy val ioCanOpenHomeMenu: IO CanOpen HomeMenu.type = _.open
-    implicit lazy val ioCanOpenHomeConfirmMenu: IO CanOpen HomeMenu.HomeRemoveConfirmationMenu =
+    implicit lazy val ioCanOpenHomeMenu: IO CanOpen home.HomeMenu = _.open
+    implicit lazy val ioCanOpenHomeConfirmMenu: IO CanOpen home.HomeRemoveConfirmationMenu =
       _.open
     implicit lazy val ioCanOpenPassiveSkillMenu: IO CanOpen PassiveSkillMenu.type = _.open
