+++ src/main/scala/com/github/unchama/seichiassist/menus/stickmenu/FirstPage.scala
@@ -15,15 +15,11 @@ import com.github.unchama.seichiassist.data.descrptions.PlayerStatsLoreGenerator
 import com.github.unchama.seichiassist.data.MenuInventoryData
 import com.github.unchama.seichiassist.effects.player.CommonSoundEffects
 import com.github.unchama.seichiassist.menus.achievement.AchievementMenu
+import com.github.unchama.seichiassist.menus.home.HomeMenu
 import com.github.unchama.seichiassist.menus.minestack.MineStackMainMenu
 import com.github.unchama.seichiassist.menus.ranking.RankingRootMenu
 import com.github.unchama.seichiassist.menus.skill.{ActiveSkillMenu, PassiveSkillMenu}
-import com.github.unchama.seichiassist.menus.{
-  CommonButtons,
-  HomeMenu,
-  RegionMenu,
-  ServerSwitchMenu
-}
+import com.github.unchama.seichiassist.menus.{CommonButtons, RegionMenu, ServerSwitchMenu}
 import com.github.unchama.seichiassist.subsystems.anywhereender.AnywhereEnderChestAPI
 import com.github.unchama.seichiassist.subsystems.anywhereender.domain.AccessDenialReason
 import com.github.unchama.seichiassist.subsystems.breakcount.BreakCountReadAPI
@@ -86,7 +82,7 @@ object FirstPage extends Menu {
     val ioCanOpenActiveSkillMenu: IO CanOpen ActiveSkillMenu.type,
     val ioCanOpenServerSwitchMenu: IO CanOpen ServerSwitchMenu.type,
     val ioCanOpenAchievementMenu: IO CanOpen AchievementMenu.type,
-    val ioCanOpenHomeMenu: IO CanOpen HomeMenu.type,
+    val ioCanOpenHomeMenu: IO CanOpen HomeMenu,
     val ioCanOpenPassiveSkillMenu: IO CanOpen PassiveSkillMenu.type,
     val ioCanOpenRankingRootMenu: IO CanOpen RankingRootMenu.type,
     val enderChestAccessApi: AnywhereEnderChestAPI[IO],
@@ -640,7 +636,7 @@ object FirstPage extends Menu {
       )
     }
 
-    def homePointMenuButton(implicit ioCanOpenHomeMenu: IO CanOpen HomeMenu.type): Button = {
+    def homePointMenuButton(implicit ioCanOpenHomeMenu: IO CanOpen HomeMenu): Button = {
       val iconItemStack =
         new IconItemStackBuilder(Material.BED)
           .title(s"$YELLOW$UNDERLINE${BOLD}ホームメニューを開く")
@@ -651,7 +647,7 @@ object FirstPage extends Menu {
         iconItemStack,
         LeftClickButtonEffect(
           FocusedSoundEffect(Sound.BLOCK_CHEST_OPEN, 1.0f, 1.5f),
-          ioCanOpenHomeMenu.open(HomeMenu)
+          ioCanOpenHomeMenu.open(new HomeMenu)
         )
       )
     }
